Extract class PracticeExercise
