+++ contribution/generator/src/Command/CreateTestsCommand.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace App\Command;
 
+use App\TrackData\PracticeExercise;
 use PhpParser\BuilderFactory;
 use PhpParser\Node;
 use PhpParser\Node\Stmt\Namespace_;
 use PhpParser\PrettyPrinter;
-use RuntimeException;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -23,19 +23,11 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 class CreateTestsCommand extends Command
 {
     private string $trackRoot;
-    private string $pathToConfiglet = '';
-    private string $pathToPracticeExercises = '';
-    private string $pathToPracticeExercise = '';
-    private string $pathToCanonicalData = '';
-
-    private string $exerciseSlug;
 
     public function __construct(
         private string $projectDir,
     ) {
         $this->trackRoot = realpath($projectDir . '/../..');
-        $this->pathToConfiglet = $this->trackRoot . '/bin/configlet';
-        $this->pathToPracticeExercises = $this->trackRoot . '/exercises/practice/';
 
         parent::__construct();
     }
@@ -45,98 +37,20 @@ class CreateTestsCommand extends Command
         $this->addArgument('exercise', InputArgument::REQUIRED, 'Exercise slug');
     }
 
-    /**
-     * @throws RuntimeException
-     */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $this->exerciseSlug = $input->getArgument('exercise');
-        $this->pathToPracticeExercise =
-            $this->pathToPracticeExercises . $this->exerciseSlug;
-
-        $this->ensureConfigletCanBeUsed();
-        $this->ensurePracticeExerciseCanBeUsed();
-        $this->pathToCachedCanonicalDataFromConfiglet();
-        $this->ensurePathToCanonicalDataCanBeUsed();
+        // TODO: Move things to `TestGenerator`
+        $exerciseSlug = $input->getArgument('exercise');
+        $exercise = new PracticeExercise(
+            $this->trackRoot,
+            $exerciseSlug,
+        );
 
         $io = new SymfonyStyle($input, $output);
-        $io->writeln('Project dir: ' . $this->projectDir);
-        $io->writeln('Generating tests for ' . $this->exerciseSlug . ' in ' . $this->pathToPracticeExercise);
-        $io->writeln('Constructed path to canonical data: ' . $this->pathToCanonicalData);
+        $io->writeln('Generating tests for ' . $exerciseSlug . ' in ' . $exercise->pathToExercise());
 
         $io->success('Generating Tests - Finished');
         return Command::SUCCESS;
-    }
-
-    private function ensurePathToCanonicalDataCanBeUsed(): void
-    {
-        if (
-            !(
-                is_readable($this->pathToCanonicalData)
-                && is_file($this->pathToCanonicalData)
-            )
-        ) {
-            throw new RuntimeException(
-                'Cannot read "configlet" provided cached canonical data from '
-                . $this->pathToCanonicalData
-                .'. Check exercise slug or access rights!'
-            );
-        }
-    }
-
-    private function ensurePracticeExerciseCanBeUsed(): void
-    {
-        if (
-            !(
-                is_writable($this->pathToPracticeExercise)
-                && is_dir($this->pathToPracticeExercise)
-            )
-        ) {
-            throw new RuntimeException(
-                'Cannot write to exercise directory. Create exercise with'
-                . ' configlet first or check access rights!'
-            );
-        }
-    }
-
-    private function ensureConfigletCanBeUsed(): void
-    {
-        if (
-            !(
-                is_executable($this->pathToConfiglet)
-                && is_file($this->pathToConfiglet)
-            )
-        ) {
-            throw new RuntimeException(
-                'configlet not found. Run the generator from track root.'
-                . ' Fetch configlet and create exercise with configlet first!'
-            );
-        }
-    }
-
-    private function pathToCachedCanonicalDataFromConfiglet(): void
-    {
-        $command = 'bash -c \'set -eo pipefail; '
-            . $this->pathToConfiglet
-            . ' -v d -t '
-            . $this->trackRoot
-            . ' info -o | head -1 | cut -d " " -f 5\''
-            ;
-        $resultCode = Command::FAILURE;
-
-        $configletCache = \exec(command: $command, result_code: $resultCode);
-        if ($configletCache === false || $resultCode !== Command::SUCCESS) {
-            throw new RuntimeException(
-                '"configlet" could not provide cached canonical data.'
-                . ' Create exercise with configlet first!'
-            );
-        }
-
-        $this->pathToCanonicalData = \sprintf(
-            '%1$s/exercises/%2$s/canonical-data.json',
-            $configletCache,
-            $this->exerciseSlug
-        );
     }
 
     // TODO: Move to own class
